Move PlayerName to DataType package
