+++ src/main/java/com/pm/server/controller/PacmanController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.pm.server.datatype.Coordinate;
 import com.pm.server.datatype.CoordinateImpl;
+import com.pm.server.datatype.PlayerName;
 import com.pm.server.datatype.PlayerState;
 import com.pm.server.exceptionhttp.BadRequestException;
 import com.pm.server.exceptionhttp.ConflictException;
@@ -21,7 +22,6 @@ import com.pm.server.exceptionhttp.InternalServerErrorException;
 import com.pm.server.exceptionhttp.NotFoundException;
 import com.pm.server.player.Pacman;
 import com.pm.server.player.PacmanImpl;
-import com.pm.server.player.PlayerName;
 import com.pm.server.repository.PacmanRepository;
 import com.pm.server.request.PlayerStateRequest;
 import com.pm.server.response.LocationResponse;
