+++ src/main/java/com/pm/server/datatype/PlayerName.java
@@ -1,4 +1,4 @@
-package com.pm.server.player;
+package com.pm.server.datatype;
 
 public enum PlayerName {
