+++ src/main/java/com/pm/server/player/GhostImpl.java
@@ -4,6 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Component;
 
+import com.pm.server.datatype.PlayerName;
 import com.pm.server.datatype.PlayerState;
 
 @Component
