+++ src/main/java/com/pm/server/player/PacmanImpl.java
@@ -2,6 +2,8 @@ package com.pm.server.player;
 
 import org.springframework.stereotype.Component;
 
+import com.pm.server.datatype.PlayerName;
+
 @Component
 public class PacmanImpl extends PlayerImpl implements Pacman {
